Fix test assuming sizeof(int) >= 4 when performing a shift
gcc/testsuite/ChangeLog:

	* gcc.c-torture/execute/20071211-1.c: Fix test to ensure
	that a shift doesn't overflow when sizeof(int) < 4

+++ gcc/testsuite/gcc.c-torture/execute/20071211-1.c
@@ -12,8 +12,8 @@ struct s
 
 int main()
 {
-  int f2;
-  sv.f2 = (1 << 24) - 1;
+  long f2;
+  sv.f2 = ((long)1 << 24) - 1;
   __asm__ volatile ("" : : : "memory");
   ++sv.f2;
   f2 = sv.f2;
